moved all pokemon handler related functions to new module
