+++ src/main.js
@@ -2,8 +2,7 @@
 /* eslint-disable no-use-before-define */
 /// <reference types="jquery" />
 import * as request from './apiRequests.js';
-import easterEgg from './easterEgg.js';
-import { hideError } from './error.js';
+import getPokemonHandler from './pokemonHandler.js';
 
 document.querySelector('#random-pokemon-button').addEventListener('click', () => {
   const random = generateRandomId();
@@ -57,29 +56,7 @@ function updateTotalPokemon(total) {
 function updateFirstPokemonIndex(offset) {
   $('#pokemon-now-showing').attr('data-index', offset);
 }
-function getPokemonHandler(pokemonName) {
-  clearPokemon();
-  addLoadingToPokemon();
-  hideError();
-  if (pokemonName === 'michelin') {
-    const pokemon = easterEgg();
-    pokemonHandler(pokemon);
-  } else {
-    request.getPokemonByIdOrName(pokemonName).then((pokemon) => {
-      if (!pokemon) {
-        return;
-      }
-      pokemonHandler(pokemon);
-    });
-  }
-}
-function addLoadingToPokemon() {
-  $('#pokemon-carousel .carousel-inner')
-    .append('<div class="loader pokemon-carousel active"></div>');
-}
-function removeLoadingFromPokemon() {
-  $('#pokemon-carousel .loader.pokemon-carousel').remove();
-}
+
 function getFirstPokemonOnListId() {
   const offset = Number($('#pokemon-now-showing').attr('data-index'));
   return offset + 1;// offset starts in 0, pokemon list starts in 1
@@ -137,49 +114,6 @@ function clearPokemonList() {
   list.children().remove();
 }
 
-function pokemonHandler(pokemon) {
-  showPokemonInfo();
-  removeLoadingFromPokemon();
-  updateName(pokemon);
-  updateTypes(pokemon.types);
-  updateAbilities(pokemon.abilities);
-  updateMoves(pokemon.moves);
-  displayPokemon(pokemon.sprites.other);
-}
-function showPokemonInfo() {
-  $('#pokemon-info').removeClass('hidden');
-}
-function updateTypes(types) {
-  types.forEach((type) => {
-    const div = $(`<div><i class="type-${type.type.name}">${type.type.name}</i></div>`);
-    $('#pokemon-types').append(div);
-  });
-}
-function updateAbilities(abilities) {
-  abilities.forEach((ability) => {
-    const div = $(`<div><i class="pokemon-ability">${ability.ability.name}</i></div>`);
-    $('#pokemon-abilities').append(div);
-  });
-}
-function updateMoves(moves) {
-  const COMPLETE_PAIR = 2;
-  let currentRow = $('<tr></tr>');
-  $('#pokemon-moves-table').append(currentRow);
-  moves.forEach((move) => {
-    const td = $(`<td style="width: 60%;">${move.move.name}</td>`);
-    if (currentRow.children().length === COMPLETE_PAIR) {
-      currentRow = $('<tr></tr>');
-      $('#pokemon-moves-table').append(currentRow);
-    }
-    currentRow.append(td);
-  });
-}
-function updateName(pokemon) {
-  const { name } = pokemon;
-  const capitalizedName = name.charAt(0).toUpperCase() + name.slice(1);
-  const nameElement = $('#pokemon-name');
-  nameElement.text(capitalizedName).addClass(`type-${pokemon.types[0].type.name}`);
-}
 function displayPokemon(sprites) {
   Object.keys(sprites).forEach((spriteCategory) => {
     Object.keys(sprites[spriteCategory]).forEach((sprite) => {
